requirements.txt
asgiref==3.5.0
Django==3.2.0
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pygments==2.11.2
pytz==2022.1
sqlparse==0.4.2
